third_party/WebKit/LayoutTests/fast/performance/performance-first-paint-timing-observable.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <script>

    async_test(function (t) {
        var observer = new PerformanceObserver(
            t.step_func(function (entryList) {
                var entries = entryList.getEntries();
                // Nothing contentful to be painted yet.
                assert_equals(entries.length, 1,
                    "There should be only first paint timing instance.");
                assert_equals(entries[0].entryType, "paint",
                    "Expected entryType to be: paint.");
                assert_equals(entries[0].name, "first-paint",
                    "Expected name to be: first-paint.");
                assert_equals(entries[0].duration, 0,
                    "Expected duration to be: 0.");
                observer.disconnect();
                t.done();
            })
        );
        observer.observe({entryTypes: ["paint"]});

    }, "Performance first paint timing entry is observable.");
    </script>
    
    <div style="background-color:black;color:white;padding:20px;"></div>
</body>
</html>
